legacy/requirements.txt
# AI Job Matcher - Python Dependencies
# Install with: pip install -r requirements.txt

# Database
psycopg2-binary>=2.9.9

# AI
openai>=1.0.0

# Environment
python-dotenv>=1.0.0

# Google APIs (for Google Docs resume upload)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Excel export
openpyxl>=3.1.2
pandas>=2.0.0

# Streamlit Dashboard (real-time monitoring)
streamlit>=1.30.0
plotly>=5.18.0

# Telegram Notifications
requests>=2.31.0

# Google Gemini AI (used by multi_ai_analyzer.py, premium_cv_generator.py)
google-generativeai>=0.3.0

# LinkedIn Job Fetcher (job_matcher.py)
beautifulsoup4>=4.12.0

# Premium CV PDF Generation (premium_cv_generator.py)
pdfkit>=1.0.0
